# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Financial Constraint Analysis Platform API
  description: Enterprise-grade financial analysis platform with AI-powered insights, constraint management, and comprehensive compliance features.
  version: 1.0.0
  contact:
    name: Financial Analyzer Team
    email: support@financial-analyzer.com
  license:
    name: Proprietary
    url: https://financial-analyzer.com/license
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.financial-analyzer.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /constraints:
    get:
      summary: List user constraints
      tags: [Constraints]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Constraints retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  constraints:
                    type: array
                    items:
                      $ref: '#/components/schemas/Constraint'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create new constraint
      tags: [Constraints]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, metric, operator, value, severity, message]
              properties:
                name:
                  type: string
                description:
                  type: string
                metric:
                  type: string
                operator:
                  type: string
                  enum: [LESS_THAN, GREATER_THAN, EQUAL, LESS_THAN_OR_EQUAL, GREATER_THAN_OR_EQUAL, NOT_EQUAL]
                value:
                  type: number
                severity:
                  type: string
                  enum: [CRITICAL, WARNING, INFO]
                message:
                  type: string
      responses:
        '201':
          description: Constraint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Constraint'

  # Analysis
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Constraint:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        metric:
          type: string
        operator:
          type: string
          enum: [LESS_THAN, GREATER_THAN, EQUAL, LESS_THAN_OR_EQUAL, GREATER_THAN_OR_EQUAL, NOT_EQUAL]
        value:
          type: number
        severity:
          type: string
          enum: [CRITICAL, WARNING, INFO]
        message:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...